# Conflicts:
#	wordweb-app/src/main/resources/view/html/search.html

+++ ekilex-etl/src/main/java/eki/ekilex/runner/AbstractLoaderRunner.java
@@ -5,6 +5,7 @@ import static java.util.stream.Collectors.toMap;
 import java.io.InputStream;
 import java.sql.Timestamp;
 import java.text.Normalizer;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
@@ -43,6 +44,7 @@ import eki.ekilex.data.transform.Source;
 import eki.ekilex.data.transform.Usage;
 import eki.ekilex.data.transform.UsageTranslation;
 import eki.ekilex.data.transform.Word;
+import eki.ekilex.service.ReportComposer;
 
 public abstract class AbstractLoaderRunner extends AbstractLoaderCommons implements InitializingBean {
 
@@ -52,17 +54,33 @@ public abstract class AbstractLoaderRunner extends AbstractLoaderCommons impleme
 	abstract void deleteDatasetData() throws Exception;
 	abstract void initialise() throws Exception;
 
+	private ReportComposer reportComposer;
+	protected boolean doReports;
+
+	private static final String REPORT_GUID_MISMATCH = "guid_mismatch";
+	private static final String REPORT_GUID_MAPPING_MISSING = "guid_mapping_missing";
+
+	protected static final String GUID_OWNER_DATASET_CODE = "ss1";
+	protected static final String COLLOC_OWNER_DATASET_CODE = "kol";
+	protected static final String ETYMOLOGY_OWNER_DATASET_CODE = "ety";
 	protected static final String DEFAULT_WORD_MORPH_CODE = "??";
 
-	private static final String SQL_SELECT_WORD_IDS_FOR_DATASET = "sql/select_word_ids_for_dataset.sql";
+	private static final String SQL_SELECT_WORD_IDS_FOR_DATASET_BY_LEX = "sql/select_word_ids_for_dataset_by_lexeme.sql";
+	private static final String SQL_SELECT_WORD_IDS_FOR_DATASET_BY_GUID = "sql/select_word_ids_for_dataset_by_guid.sql";
 	private static final String SQL_SELECT_MEANING_IDS_FOR_DATASET = "sql/select_meaning_ids_for_dataset.sql";
-	private static final String SQL_DELETE_DEFINITIONS_FOR_DATASET = "sql/delete_definitions_for_dataset.sql";
 	private static final String SQL_SELECT_WORD_BY_FORM_AND_HOMONYM = "sql/select_word_by_form_and_homonym.sql";
 	private static final String SQL_SELECT_WORD_BY_DATASET_AND_GUID = "sql/select_word_by_dataset_and_guid.sql";
 	private static final String SQL_SELECT_WORD_MAX_HOMONYM = "sql/select_word_max_homonym.sql";
 	private static final String SQL_SELECT_LEXEME_FREEFORM_BY_TYPE_AND_VALUE = "sql/select_lexeme_freeform_by_type_and_value.sql";
 	private static final String SQL_SELECT_SOURCE_BY_TYPE_AND_NAME = "sql/select_source_by_type_and_name.sql";
 	private static final String SQL_SELECT_WORD_GROUP_WITH_MEMBERS = "sql/select_word_group_with_members.sql";
+
+	private static final String SQL_DELETE_DEFINITIONS_FOR_DATASET = "sql/delete_definitions_for_dataset.sql";
+	private static final String SQL_DELETE_DEFINITION_FF_FOR_DATASET = "sql/delete_definition_freeforms_for_dataset.sql";
+	private static final String SQL_DELETE_MEANING_FF_FOR_DATASET = "sql/delete_meaning_freeforms_for_dataset.sql";
+	private static final String SQL_DELETE_COLLOCATION_FF_FOR_DATASET = "sql/delete_collocation_freeforms_for_dataset.sql";
+	private static final String SQL_DELETE_LEXEME_FF_FOR_DATASET = "sql/delete_lexeme_freeforms_for_dataset.sql";
+
 	private static final String CLASSIFIERS_MAPPING_FILE_PATH = "./fileresources/csv/classifier-main-map.csv";
 
 	private static final char[] RESERVED_DIACRITIC_CHARS = new char[] {'õ', 'ä', 'ö', 'ü', 'š', 'ž', 'Õ', 'Ä', 'Ö', 'Ü', 'Š', 'Ž'};
@@ -83,9 +101,9 @@ public abstract class AbstractLoaderRunner extends AbstractLoaderCommons impleme
 	protected static final String EKI_CLASSIIFER_ETYMKEELTYYP = "etymkeel_tyyp";
 	protected static final String EKI_CLASSIFIER_ENTRY_CLASS = "entry class";
 
-	private String sqlSelectWordIdsForDataset;
+	private String sqlSelectWordIdsForDatasetByLexeme;
+	private String sqlSelectWordIdsForDatasetByGuid;
 	private String sqlSelectMeaningIdsForDataset;
-	private String sqlDeleteDefinitionsForDataset;
 	private String sqlSelectWordByFormAndHomonym;
 	private String sqlSelectWordByDatasetAndGuid;
 	private String sqlSelectWordMaxHomonym;
@@ -93,7 +111,16 @@ public abstract class AbstractLoaderRunner extends AbstractLoaderCommons impleme
 	private String sqlSourceByTypeAndName;
 	private String sqlWordGroupWithMembers;
 
+	private String sqlDeleteDefinitionsForDataset;
+	private String sqlDeleteDefinitionFreeformsForDataset;
+	private String sqlDeleteMeaningFreeformsForDataset;
+	private String sqlDeleteCollocationFreeformsForDataset;
+	private String sqlDeleteLexemeFreeformsForDataset;
+
 	private Pattern ekiEntityPatternV;
+
+	private long t1;
+	private long t2;
 
 	@Override
 	public void afterPropertiesSet() throws Exception {
@@ -103,14 +130,14 @@ public abstract class AbstractLoaderRunner extends AbstractLoaderCommons impleme
 		ClassLoader classLoader = this.getClass().getClassLoader();
 		InputStream resourceFileInputStream;
 
-		resourceFileInputStream = classLoader.getResourceAsStream(SQL_SELECT_WORD_IDS_FOR_DATASET);
-		sqlSelectWordIdsForDataset = getContent(resourceFileInputStream);
+		resourceFileInputStream = classLoader.getResourceAsStream(SQL_SELECT_WORD_IDS_FOR_DATASET_BY_LEX);
+		sqlSelectWordIdsForDatasetByLexeme = getContent(resourceFileInputStream);
+
+		resourceFileInputStream = classLoader.getResourceAsStream(SQL_SELECT_WORD_IDS_FOR_DATASET_BY_GUID);
+		sqlSelectWordIdsForDatasetByGuid = getContent(resourceFileInputStream);
 
 		resourceFileInputStream = classLoader.getResourceAsStream(SQL_SELECT_MEANING_IDS_FOR_DATASET);
 		sqlSelectMeaningIdsForDataset = getContent(resourceFileInputStream);
-
-		resourceFileInputStream = classLoader.getResourceAsStream(SQL_DELETE_DEFINITIONS_FOR_DATASET);
-		sqlDeleteDefinitionsForDataset = getContent(resourceFileInputStream);
 
 		resourceFileInputStream = classLoader.getResourceAsStream(SQL_SELECT_WORD_BY_FORM_AND_HOMONYM);
 		sqlSelectWordByFormAndHomonym = getContent(resourceFileInputStream);
@@ -130,7 +157,38 @@ public abstract class AbstractLoaderRunner extends AbstractLoaderCommons impleme
 		resourceFileInputStream = classLoader.getResourceAsStream(SQL_SELECT_WORD_GROUP_WITH_MEMBERS);
 		sqlWordGroupWithMembers = getContent(resourceFileInputStream);
 
+		resourceFileInputStream = classLoader.getResourceAsStream(SQL_DELETE_DEFINITIONS_FOR_DATASET);
+		sqlDeleteDefinitionsForDataset = getContent(resourceFileInputStream);
+
+		resourceFileInputStream = classLoader.getResourceAsStream(SQL_DELETE_DEFINITION_FF_FOR_DATASET);
+		sqlDeleteDefinitionFreeformsForDataset = getContent(resourceFileInputStream);
+
+		resourceFileInputStream = classLoader.getResourceAsStream(SQL_DELETE_MEANING_FF_FOR_DATASET);
+		sqlDeleteMeaningFreeformsForDataset = getContent(resourceFileInputStream);
+
+		resourceFileInputStream = classLoader.getResourceAsStream(SQL_DELETE_COLLOCATION_FF_FOR_DATASET);
+		sqlDeleteCollocationFreeformsForDataset = getContent(resourceFileInputStream);
+
+		resourceFileInputStream = classLoader.getResourceAsStream(SQL_DELETE_LEXEME_FF_FOR_DATASET);
+		sqlDeleteLexemeFreeformsForDataset = getContent(resourceFileInputStream);
+
 		ekiEntityPatternV = Pattern.compile("(&(ehk|Hrl|hrl|ja|jne|jt|ka|nt|puudub|v|vm|vms|vrd|vt|напр.|и др.|и т. п.|г.);)");
+	}
+
+	protected void start() throws Exception {
+		if (doReports) {
+			reportComposer = new ReportComposer(getDataset() + " unified loader", REPORT_GUID_MISMATCH, REPORT_GUID_MAPPING_MISSING);
+		}
+		logger.debug("Loading \"{}\" ...", getDataset());
+		t1 = System.currentTimeMillis();
+	}
+
+	protected void end() throws Exception {
+		t2 = System.currentTimeMillis();
+		logger.debug("Done loading \"{}\" in {} ms", getDataset(), (t2 - t1));
+		if (reportComposer != null) {
+			reportComposer.end();
+		}
 	}
 
 	protected boolean isLang(String lang) {
@@ -198,30 +256,54 @@ public abstract class AbstractLoaderRunner extends AbstractLoaderCommons impleme
 		return cleanValue;
 	}
 
-	//TODO under construction
 	protected void deleteDatasetData(String dataset) throws Exception {
 
 		Map<String, Object> tableRowParamMap = new HashMap<>();
 		tableRowParamMap.put("dataset", dataset);
 
-		List<Long> wordIds = basicDbService.queryList(sqlSelectWordIdsForDataset, tableRowParamMap, Long.class);
-		logger.debug("There are {} words in \"{}\" to be deleted", wordIds.size(), dataset);
+		List<Long> wordIdsLex = basicDbService.queryList(sqlSelectWordIdsForDatasetByLexeme, tableRowParamMap, Long.class);
+		List<Long> wordIdsGuid = basicDbService.queryList(sqlSelectWordIdsForDatasetByGuid, tableRowParamMap, Long.class);
+		List<Long> wordIds = new ArrayList<>();
+		wordIds.addAll(wordIdsLex);
+		wordIds.addAll(wordIdsGuid);
+		wordIds = wordIds.stream().distinct().collect(Collectors.toList());
+		logger.debug("There are {} words in \"{}\" to be deleted - {} by lexemes, {} by guids", wordIds.size(), dataset, wordIdsLex.size(), wordIdsGuid.size());
 
 		List<Long> meaningIds = basicDbService.queryList(sqlSelectMeaningIdsForDataset, tableRowParamMap, Long.class);
 		logger.debug("There are {} meanings in \"{}\" to be deleted", meaningIds.size(), dataset);
 
 		String sql;
 
+		// freeforms
+		basicDbService.executeScript(sqlDeleteDefinitionFreeformsForDataset, tableRowParamMap);
+		basicDbService.executeScript(sqlDeleteMeaningFreeformsForDataset, tableRowParamMap);
+		basicDbService.executeScript(sqlDeleteLexemeFreeformsForDataset, tableRowParamMap);
+
+		// delete definitions
+		basicDbService.executeScript(sqlDeleteDefinitionsForDataset, tableRowParamMap);
+
+		// delete collocations + freeforms
+		if (StringUtils.equals(COLLOC_OWNER_DATASET_CODE, dataset)) {
+			basicDbService.executeScript(sqlDeleteCollocationFreeformsForDataset);
+			sql = "delete from " + COLLOCATION;
+			basicDbService.executeScript(sql);
+		}
+
+		// delete etymology
+		if (StringUtils.equals(ETYMOLOGY_OWNER_DATASET_CODE, dataset)) {
+			sql = "delete from " + WORD_ETYMOLOGY;
+			basicDbService.executeScript(sql);
+		}
+
 		// delete lexemes
 		sql = "delete from " + LEXEME + " l where l.dataset_code = :dataset";
 		basicDbService.executeScript(sql, tableRowParamMap);
 
 		// delete word guids
-		sql = "delete from " + WORD_GUID + " wg where wg.dataset_code = :dataset";
-		basicDbService.executeScript(sql, tableRowParamMap);
-
-		// delete definitions
-		basicDbService.executeScript(sqlDeleteDefinitionsForDataset, tableRowParamMap);
+		if (!StringUtils.equals(GUID_OWNER_DATASET_CODE, dataset)) {
+			sql = "delete from " + WORD_GUID + " wg where wg.dataset_code = :dataset";
+			basicDbService.executeScript(sql, tableRowParamMap);
+		}
 
 		// delete words
 		sql = "delete from " + WORD + " where id = :wordId";
@@ -331,6 +413,7 @@ public abstract class AbstractLoaderRunner extends AbstractLoaderCommons impleme
 
 		List<Guid> mappedGuids = ssGuidMap.get(guid);
 		if (CollectionUtils.isEmpty(mappedGuids)) {
+			appendToReport(REPORT_GUID_MAPPING_MISSING, dataset, wordValue, guid);
 			return createOrSelectWord(word, paradigms, dataset, reusedWordCount);
 		}
 
@@ -338,16 +421,15 @@ public abstract class AbstractLoaderRunner extends AbstractLoaderCommons impleme
 
 			String ssWordValue = ssGuidObj.getWord();
 			String ssGuid = ssGuidObj.getValue();
-			String ssDataset = "ss1";
 
 			if (StringUtils.equalsIgnoreCase(wordValue, ssWordValue)) {
-				List<Map<String, Object>> tableRowValueMaps = getWord(wordValue, ssGuid, ssDataset);
+				List<Map<String, Object>> tableRowValueMaps = getWord(wordValue, ssGuid, GUID_OWNER_DATASET_CODE);
 				Map<String, Object> tableRowValueMap = null;
 				if (CollectionUtils.size(tableRowValueMaps) == 1) {
 					tableRowValueMap = tableRowValueMaps.get(0);
 				} else if (CollectionUtils.size(tableRowValueMaps) > 1) {
 					tableRowValueMap = tableRowValueMaps.get(0);
-					logger.warn("There are multiple words with same value and guid in {}: \"{}\" - \"{}\"", ssDataset, wordValue, ssGuid);
+					logger.warn("There are multiple words with same value and guid in {}: \"{}\" - \"{}\"", GUID_OWNER_DATASET_CODE, wordValue, ssGuid);
 				}
 				if (tableRowValueMap == null) {
 					return createOrSelectWord(word, paradigms, dataset, reusedWordCount);
@@ -367,6 +449,7 @@ public abstract class AbstractLoaderRunner extends AbstractLoaderCommons impleme
 		}
 		List<String> mappedWordValues = mappedGuids.stream().map(Guid::getWord).collect(Collectors.toList());
 		logger.debug("Word value doesn't match guid mapping(s): \"{}\" / \"{}\"", wordValue, mappedWordValues);
+		appendToReport(REPORT_GUID_MISMATCH, dataset, wordValue, guid, mappedWordValues, "Sõnad ei kattu");
 
 		return createOrSelectWord(word, paradigms, dataset, reusedWordCount);
 	}
@@ -1081,4 +1164,11 @@ public abstract class AbstractLoaderRunner extends AbstractLoaderCommons impleme
 				.collect(toMap(cells -> cells[2], cells -> cells[6], (c1, c2) -> c2));
 	}
 
+	private void appendToReport(String reportName, Object ... reportCells) throws Exception {
+		if (!doReports) {
+			return;
+		}
+		String logRow = StringUtils.join(reportCells, CSV_SEPARATOR);
+		reportComposer.append(reportName, logRow);
+	}
 }
